Makefile for docker image build
This builds in two stages, to make a minimal image:

 - run a golang tools image with the source to produce the executable
   file;
 - build an image with the executable file and iptables

I use an alpine linux image: the scratch image doesn't have iptables,
and it is easier to start from alpine than to create a
(statically-linked) iptables.

A subdirectory `build/` in which to put the results of `go get` is
also mounted into the build image. This speeds up the build by keeping
dependencies around.