Use container and script for build
Since contained processes run as root, any files created will be owned
by root. One can tell the container to run as a specific UID, but it
would have to be a user first.

In addition, we want to create a binary that will run in Alpine, which
means a static binary. To create a static binary, we have to have
CGO_ENABLED=0 -- and that will in general provoke go into rebuilding
its standard library.

So: we make a build container first, with a CGO_ENABLED=0 standard
library. Then, to actually build the exe, we run a small script which
creates a user with the same UID as the mounted directory and `su`s to
that user so that files end up being owned by the same user. Phew!